requirements.txt
inquirer==2.7.0
PyYAML==5.3.1
